refactor: move in-app wallet layout and metadata to separate file
