+++ apps/playground-web/src/app/connect/in-app-wallet/sponsor/page.tsx
@@ -1,39 +1,13 @@
 import { CodeExample } from "@/components/code/code-example";
-import ThirdwebProvider from "@/components/thirdweb-provider";
-import { metadataBase } from "@/lib/constants";
-import type { Metadata } from "next";
-import { APIHeader } from "../../../../components/blocks/APIHeader";
 import { SponsoredInAppTxPreview } from "../../../../components/in-app-wallet/sponsored-tx";
-
-export const metadata: Metadata = {
-  metadataBase,
-  title: "Sign In, Account Abstraction and SIWE Auth | thirdweb Connect",
-  description:
-    "Let users sign up with their email, phone number, social media accounts or directly with a wallet. Seamlessly integrate account abstraction and SIWE auth.",
-};
+import ThirdwebProvider from "../../../../components/thirdweb-provider";
 
 export default function Page() {
   return (
     <ThirdwebProvider>
-      <main className="pb-20 container px-0">
-        <APIHeader
-          title="Onboard users to web3"
-          description={
-            <>
-              Onboard anyone with flexible auth options, secure account
-              recovery, and smart account integration.
-            </>
-          }
-          docsLink="https://portal.thirdweb.com/connect/in-app-wallet/overview"
-          heroLink="/in-app-wallet.png"
-        />
-
-        <section className="space-y-8">
-          <SponsoredInAppTx />
-        </section>
-
-        <div className="h-14" />
-      </main>
+      <section className="space-y-8">
+        <SponsoredInAppTx />
+      </section>
     </ThirdwebProvider>
   );
 }
